# Conflicts:
#   .idea/.gitignore
#   app/build.gradle.kts

+++ app/src/main/java/io/gitlab/fsc_clam/fscwhereswhat/model/local/NoteItem.kt
@@ -29,7 +29,7 @@ import java.net.URL
  */
 data class NoteItem(
 	val comment: String,
-	val reference: Int,
+	val reference: Long,
 	val type: EntityType,
 	val image: Image,
 	val referenceName: String
